chore(utils): refactor for more efficient webpack
