+++ packages/salesforcedx-utils-vscode/src/context/index.ts
@@ -4,4 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export { OrgInfo, WorkspaceContextUtil } from './workspaceContextUtil';
+
+/**
+ * This file is intentionally left blank. WorkspaceContextUtil and OrgInfo are exported
+ * in the root of salesforcedx-utils-vscode.
+ */
