+++ packages/salesforcedx-utils-vscode/src/context/workspaceContextUtil.ts
@@ -8,7 +8,7 @@
 import { AuthInfo, Connection } from '@salesforce/core';
 import { join } from 'path';
 import * as vscode from 'vscode';
-import { AuthUtil } from '../auth';
+import { AuthUtil } from '..';
 import { nls } from '../messages';
 import { SFDX_CONFIG_FILE, SFDX_FOLDER } from '../types';
 import { getRootWorkspacePath } from '../workspaces';
