+++ packages/system-tests/assets/sfdx-simple/force-app/main/default/classes/TestApexTrigger.trigger
@@ -1,0 +1,3 @@
+trigger TestApexTrigger on SOBJECT (before insert) {
+
+}
