+++ packages/system-tests/assets/sfdx-simple/force-app/main/default/classes/TestApexTrigger.trigger-meta.xml
@@ -1,0 +1,5 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<ApexTrigger xmlns="http://soap.sforce.com/2006/04/metadata">
+  <apiVersion>50.0</apiVersion>
+  <status>Active</status>
+</ApexTrigger>
